+++ Tests/Runtime/Scripts/Components/SimpleMessageAwareComponent.cs
@@ -51,10 +51,9 @@ namespace DxMessaging.Tests.Runtime.Scripts.Components
         private MessageRegistrationHandle? _slowComplexTargetingHandle;
         private MessageRegistrationHandle? _fastComplexTargetingHandle;
 
-        protected override bool RegisterForStringMessages => false;
-
         protected override void RegisterMessageHandlers()
         {
+            base.RegisterMessageHandlers();
             _ = _messageRegistrationToken.RegisterUntargeted<SimpleUntargetedMessage>(
                 HandleSimpleUntargetedMessage
             );